% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_primary_schedule.R
\name{assign_days}
\alias{assign_days}
\title{create the schedule as a data frame}
\usage{
assign_days(names_data, dates)
}
\arguments{
\item{names_data}{a list of kids names and their active status}

\item{dates}{dates from \code{calc_primary_sundays}}
}
\description{
create the schedule as a data frame
}
\examples{
primary <- read.csv('/Users/bridgernorman/Downloads/primary.csv')
assign_days(primary, dates)

}
