% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_primary_schedule.R
\name{calc_primary_sundays}
\alias{calc_primary_sundays}
\title{returns a dataframe with all the sundays for a certain year}
\usage{
calc_primary_sundays(year, special_sundays = list())
}
\arguments{
\item{year}{the year of dates you want as a string}

\item{special_sundays}{other special dates that might need filtered like
Christmas as a list see examples below}
}
\description{
returns a dataframe with all the sundays for a certain year
}
\examples{

calc_primary_sundays('2025')

# filtering out special dates like conference
calc_primary_sundays('2025', list('2025-01-05', '2025-01-12'))

}
