% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_largest_group.R
\name{color_largest_group}
\alias{color_largest_group}
\title{Title}
\usage{
color_largest_group(
  df,
  group_col,
  num_col,
  color1 = "steelblue",
  color2 = "gray"
)
}
\arguments{
\item{df}{the dataframe used for plotting}

\item{group_col}{the categorical value that contains the largest group}

\item{num_col}{the numeric value that defines what is the largest group}

\item{color1}{the color you want the largest bar to be}

\item{color2}{the color you want all other bars to be}
}
\value{
A scale_fill_manual argument to be used inside a ggplot
}
\description{
Title
}
\examples{
# Example usage
df <- data.frame(
  group = c("A", "B", "C", "D"),
  value = c(10, 20, 5, 15)
)

ggplot2::ggplot(df, ggplot2::aes(x = group, y = value, fill = group)) +
  ggplot2::geom_col() +
  color_largest_group(df, "group", "value")

}
