% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_holiday.R
\name{is_holiday}
\alias{is_holiday}
\title{is_holiday}
\usage{
is_holiday(dates)
}
\arguments{
\item{date}{Any date (data type must be date)}
}
\value{
a TRUE or FALSE if the date provided was a holiday
}
\description{
Takes in a date and return true if it is a holiday specified by SGH
}
